backend/requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
pypdf>=3.15.1
openai>=1.1.1
python-dotenv>=1.0.0
tiktoken>=0.5.1
chromadb>=0.4.13
pydantic>=2.4.2
